refactor(strategies): decouple API from signal checks
